api/requirements.txt
Flask==3.1.3
Flask-Cors==4.0.0
python-dotenv==1.0.1